🔧 Fix 3 errores de compilación Kotlin (run #4)
- StatsRepository: no se puede asignar el class object de un enum a una variable -> uso directo de Achievement.*
- ParticleView: emojis como Array<out String> y setModeEmoji asigna el vararg directo
- RouletteView: duración del giro decayó a Int -> .toLong()

+++ app/src/main/java/com/studiolexair/preguntascalientes/utils/views/RouletteView.kt
@@ -64,7 +64,7 @@ class RouletteView @JvmOverloads constructor(
 
         val start = 0f
         ValueAnimator.ofFloat(0f, finalAngle).apply {
-            duration = 3400 + Random.nextInt(700)
+            duration = (3400 + Random.nextInt(700)).toLong()
             interpolator = DecelerateInterpolator(1.6f)
             addUpdateListener {
                 currentAngle = it.animatedValue as Float
